Add the lazy load functionality and use more jquery-ui.
This is a back commit of work I've done on making this work with lazy
loading.  The framework now exists to talk to a backend server to
retrieve the nodes of the tree on demand.

In the process, I added the possibility to use more jquery-ui elements,
localized the options for a TreeTable, and moved the public functions
into the jquery-ui best practice of calling it through the main function
by name.  So intead of

$('tree').expand();

it's now

$('tree').treeTable("expand");